chore: rename directory correctly
